Return default value instead of False when all clients are down
The documentation for `ignore_exc` says it will treat memcache/network
errors as cache misses (i.e. `None` for reads and `False` for writes).

However the short-circuit path for when no client can be used (because
they're all marked as down/bad) always returns `False`. This change
change the behaviour to return each command's default value instead.

I actually stumbled upon this because I wanted to inspect some behaviour
in a system without cache so I stopped all memcached nodes. Our code
looked something like this:

    data = memcached_client.get(key)
    if data is None:
        return fetch_from_database(key)
    else:
        return json.loads(data)

What happened was that every once in a while I would see a stacktrace
because the `json.loads(data)` call would fail with a `TypeError`
because it tries to deserialize a bool (`False`).