Always close the existing socket on _connect()
I don't currently see a way in which we would enter _connect() with an
existing socket, but let's explicitly close() first just to be safe.
This also ensures we enter the rest of the _connect() code knowing that
`self.sock is None`.

While I'm here, add better test coverage for the close() method.