Always send command keys in their original order
Some commands are sensitive to the order of their key arguments so
revise the previous code to explicitly build a list of prefixed keys
and a dictionary of remapped keys. This generalization also lets us
remove command-specific checks from this code path.

The list of prefixed keys is used to build the command string
(preserving order).

The dictionary of remapped keys is used to map the prefixed keys from
the response to the key names that were originally provided to the
client.