Parse version directly from pymemcache/__init__.py
We can no longer import the __version__ attribute because there is no an
implicit runtime dependency on six (as of #197), and we can't guarantee
that six is installed until *after* setup.py is parsed and run.

Instead, parse the `__version__ = 'x.y.z'` string from __init__.py to
extract the version.

Fixes #214