Check integer input values to Client methods
If a caller tries to send
- a fractional expiration time
- a string for the "incr" command
- delay=None (instead of delay=0)

The pymemcache base client will now catch these usage mistakes a
throw a MemcacheIllegalInputError.

Although there are existing valid (but undocumented) usages of the
current string conversion behavior, the docs state that these values
are supposed to be ints. Now if a non-int value is used, it will
immediately trigger an error.

This behavior is superior to the prior behavior especially in the
case where a command is used with `noreply=True`. It can fail to
parse, but the client won't pick up on this until the next call to
memcache which reads the reply.

The especially bad case we're seeking to avoid is this:

    from pymemcache.client.base import Client
    client = Client(("localhost", 11211))
    client.set("foo", "bar", expire=1.5, noreply=True)
    client.get("foo")  # triggers MemcacheUnknownCommandError

Applies to all commands using `expire`, `delay`, incr and decr
inputs, and memlimit.

Because this is can be a breaking change for some users, the changelog
entry is written under 3.0.0 (unreleased).

closes #240