Add check that `expire` is an integer
If a caller tries to send a fractional expiration time, that should be
caught and rejected with a MemcacheIllegalInputError. Otherwise, calls
with `noreply=True` could fail to parse, but the client won't pick up on
this until the next call to memcache which reads the reply.

The case we're seeking to avoid is this:

    from pymemcache.client.base import Client
    client = Client(("localhost", 11211))
    client.set("foo", "bar", expire=1.5, noreply=True)
    client.get("foo")  # triggers MemcacheUnknownCommandError

In addition to allowing for values which are `int` (or `long` on py2),
this implementation allows for `expire` values to be a text type (str
on py2, bytes or str on py3), so long as `int(expire)` correctly
converts it. If `int(expire)` fails (ValueError), that indicates that
the input could not be treated as an int and is a failure.

Applies `Client._store_cmd`, which covers a class of commands, and
`Client.touch`, which appears to be the only command with a non-shared
implementation that includes `expire` as a parameter.