Be more consistent about using Client.encoding (#242)
This is a stylistic issue with little chance for actual impact. Some
methods use 'ascii' verbatim and some use `self.encoding` (which
defaults to 'ascii'). Those which are using `ascii` are only supposed to
be accepting integer inputs, so this is actually fine.

However, the result is a potentially confusing discrepancy -- why does
`Client.touch` use `ascii` while `Client.add` uses `self.encoding`?
To alleviate this potential confusion, be consistent about applying
`self.encoding` to all memcached commands in the base client.

This does not impact the `_check_key` func, which re-encodes potential
unicode inputs as ASCII.