Switch to GitHub Actions (#252)
GitHub Actions profiles a nicer alternate to Travis for running
continuous integration jobs.

One compromise I've made in this migration is moving away from
using `tox` to run tests in our CI environment. We were previously
relying on the tox-travis to bridge between Travis' build matrix
and tox's environment list. At the moment, there doesn't appear to
be an equivalent for GitHub actions. Once there is one, we can
consider switching to it so that local and CI uses the same test
runner.