Remove idle connections from pool
The Memcached server can be configured to drop idle connections (via the
`idle_timeout` option). When this occurs, pymemcache may still try to
use the connection, resulting in `MemcacheUnexpectedCloseError`. Hence
the need for client-side idle timeout logic.

This change was proposed in github.com/pinterest/pymemcache#114, but
ultimately dropped. The current PR uses the same approach but differs
slightly in the details.