Disable IPv6 localhost resolution for CI
GitHub Actions appears to be busted with regard to IPv6-based localhost
resolution between containers. This prevents us from successfully
connecting to our memcached service containers.

We can't resolve this by using `--server` (and `--tls-server`) to
explicitly specify 127.0.0.1 to force IPv4, but that unfortunately
breaks TLS certification verification because that expects 'localhost'.

We should come up with a better solution for this, but for the short
term, this unblocks the failing CI builds.