[WIP] Adding retry management
Still work in progress.

The goal is to decorate all passed client dynamically to avoid technical
debt and API changes.

This module will only decorate passed client to manage uniformized and
standardized retries. It will be possible to configure the retry
strategies and mechanismes.

Here is just a PoC for now.