Ensure us to close socket on MemcacheUnexpectedCloseError
During network maintenance operations or during server reboot users could
face MemcacheUnexpectedCloseError.

This kind of exception will be raised down the stack. It's a design
choice [1]. Users should manage this kind of exception.

However, at some points the opened sockets remains more or less opened
after this kind of fail on the client side [2]. It will require n+1
calls to see them recreated when the connection will be possible. That's
could be problem with services that implement several cache logics in parallel
and where their clients are based on pymemcache.

These changes simply ensure to close the socket properly. The socket
will be recreated by the client if needed.

[1] https://github.com/pinterest/pymemcache/issues/307
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1977711