[py27] Fixing a bug with bad conditions
These changes fix a condition for python 2.7 only.

Python 3 and python 2.7 behave differently during our key handling.

With python 3, in the case where the key is empty and where the key is not a
bytes object the concatenation [1] will fail and the execution will be stoped
with the raise of a TypeError exception.

With python 2 the concatenation is not a problem and so the execution
isn't stoped and that lead us to a problematic condition few line below [2].
Here we execute a condition with a boolean logic and an `or` operator.
That mean that if the first condition doesn't return `True` we will
execute the second condition, in this case we will compare an index of a
previous split with our original key. The problem here is that if our
key is empty then we will try to retrieve the index `0` which will lead
us to an IndexError.

This error is allowed with python 2.7 because the previous concatenation
doesn't fail. The execution can continue.

These changes ensure that the list is not empty before trying to
retrieve an index. This is possible by using an `and` operator in the
second logic expression. This one will stop the execution and will
ignore the index part if the list is empty.

These changes also add few unittest to check more directly this
function.

The empty key lead to a specific concatenation behaviour which
depends-on the version of python used and lead us to the part bug with
the condition.

[1] https://github.com/pinterest/pymemcache/blob/master/pymemcache/client/base.py#L110
[2] https://github.com/pinterest/pymemcache/blob/master/pymemcache/client/base.py#L116