Allow us to import the KeepalivOpts from the root pymemcache namespace
As this class aim to be used with all kind of clients it would be worth
to allow to import it from the root namespace of pymemcache.

Else, without these changes, we will have to import it from
`pymemcache.client.base`.

Examples without these changes:

```
>>> from pymemcache import KeepaliveOpts
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  ImportError: cannot import name 'KeepaliveOpts' from 'pymemcache'
  (/home/hberaud/pymemcache/pymemcache/__init
  __.py)
```

Examples with these changes:

```
>>> from pymemcache import KeepaliveOpts
>>>
```