Return default from hash client when using positional argument (#354)
When using HashClient with ignore_exc, get would always return
None if no server is available and the default is passed as a
positional argument. The other clients return the default
value in this case. An earlier fix only had the desired effect when
passing default as a keyword argument.

For example, Django passes the default as a positional argument.

Return the default value so HashClient behaves like the other
clients.